# backend/app/internal/app/api/openapi/openapi.yaml
openapi: 3.0.0
info:
  title: おかん予報
  version: "1.0"
  description: おかん予報
servers:
  - url: "http://localhost:8080"
paths:
  /alerts:
    get:
      summary: アラート一覧取得
      parameters:
        - schema:
            type: integer
          in: query
          name: page
          description: ページ数
        - schema:
            type: integer
          in: query
          name: limit
          description: 1ページあたりの上限数
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Alert"
    post:
      summary: アラート登録
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Alert"
      responses:
        "201":
          description: Created

  /alerts/{alert_id}:
    parameters:
      - schema:
          type: string
        name: alert_id
        in: path
        required: true

    put:
      summary: アラート変更
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Alert"
      responses:
        "204":
          description: No Content

    delete:
      summary: アラート変更
      responses:
        "204":
          description: No Content

components:
  schemas:
    Alert:
      properties:
        id:
          type: string
          example: 01F7ZF7XC2CC45NPTD8TMFFV4A
        title:
          type: string
          example: 雨が降ったとき教えて！
        lat:
          description: 緯度
          type: number
          format: float
          example: 35.6769883
        lon:
          description: 経度
          type: number
          format: float
          example: 139.7588499
        wx:
          type: integer
          nullable: true
          enum:
            - 1
            - 2
            - 3
            - 4
            - 5
          description: >
            天気:
            `1` - 晴
            `2` - 曇
            `3` - 雨
            `4` - 雷
            `5` - 雪
        temp:
          description: 気温
          type: number
          format: float
          example: 20.4
        arpress:
          type: integer
          nullable: true
          enum:
            - 1
            - 2
            - 3
          description: >
            気圧:
            `1` - 高気圧
            `2` - 通常
            `3` - 低気圧
        wndspd:
          type: integer
          nullable: true
          enum:
            - 1
            - 2
            - 3
          description: >
            風速:
            `1` - ？
            `2` - ？
            `3` - ？
        rhum:
          type: integer
          nullable: true
          enum:
            - 1
            - 2
            - 3
          description: >
            湿度:
            `1` - ？
            `2` - ？
            `3` - ？
        feeltmp:
          type: integer
          nullable: true
          enum:
            - 1
            - 2
            - 3
          description: >
            体感気温:
            `1` - ？
            `2` - ？
            `3` - ？
        text:
          description: 通知内容
          type: string
          example: おはよう！今日は雨が降っているよ。傘を持つようにしよう！
      required:
        - id
        - title
        - lat
        - lon
        - text

  securitySchemes:
    Bearer:
      type: http
      scheme: bearer