# backend/app/internal/app/api/openapi/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.8.2 DO NOT EDIT.
package openapi

// Defines values for AlertArpress.
const (
	AlertArpressN1 AlertArpress = 1

	AlertArpressN2 AlertArpress = 2

	AlertArpressN3 AlertArpress = 3
)

// Defines values for AlertFeeltmp.
const (
	AlertFeeltmpN1 AlertFeeltmp = 1

	AlertFeeltmpN2 AlertFeeltmp = 2

	AlertFeeltmpN3 AlertFeeltmp = 3
)

// Defines values for AlertRhum.
const (
	AlertRhumN1 AlertRhum = 1

	AlertRhumN2 AlertRhum = 2

	AlertRhumN3 AlertRhum = 3
)

// Defines values for AlertWndspd.
const (
	AlertWndspdN1 AlertWndspd = 1

	AlertWndspdN2 AlertWndspd = 2

	AlertWndspdN3 AlertWndspd = 3
)

// Defines values for AlertWx.
const (
	AlertWxN1 AlertWx = 1

	AlertWxN2 AlertWx = 2

	AlertWxN3 AlertWx = 3

	AlertWxN4 AlertWx = 4

	AlertWxN5 AlertWx = 5
)

// Alert defines model for Alert.
type Alert struct {
	// 気圧: `1` - 高気圧 `2` - 通常 `3` - 低気圧
	Arpress *AlertArpress `json:"arpress"`

	// 体感気温: `1` - ？ `2` - ？ `3` - ？
	Feeltmp *AlertFeeltmp `json:"feeltmp"`
	Id      string        `json:"id"`

	// 緯度
	Lat float32 `json:"lat"`

	// 経度
	Lon float32 `json:"lon"`

	// 湿度: `1` - ？ `2` - ？ `3` - ？
	Rhum *AlertRhum `json:"rhum"`

	// 気温
	Temp *float32 `json:"temp,omitempty"`

	// 通知内容
	Text  string `json:"text"`
	Title string `json:"title"`

	// 風速: `1` - ？ `2` - ？ `3` - ？
	Wndspd *AlertWndspd `json:"wndspd"`

	// 天気: `1` - 晴 `2` - 曇 `3` - 雨 `4` - 雷 `5` - 雪
	Wx *AlertWx `json:"wx"`
}

// 気圧: `1` - 高気圧 `2` - 通常 `3` - 低気圧
type AlertArpress int

// 体感気温: `1` - ？ `2` - ？ `3` - ？
type AlertFeeltmp int

// 湿度: `1` - ？ `2` - ？ `3` - ？
type AlertRhum int

// 風速: `1` - ？ `2` - ？ `3` - ？
type AlertWndspd int

// 天気: `1` - 晴 `2` - 曇 `3` - 雨 `4` - 雷 `5` - 雪
type AlertWx int

// GetAlertsParams defines parameters for GetAlerts.
type GetAlertsParams struct {
	// ページ数
	Page *int `json:"page,omitempty"`

	// 1ページあたりの上限数
	Limit *int `json:"limit,omitempty"`
}

// PostAlertsJSONBody defines parameters for PostAlerts.
type PostAlertsJSONBody Alert

// PutAlertsAlertIdJSONBody defines parameters for PutAlertsAlertId.
type PutAlertsAlertIdJSONBody Alert

// PostAlertsJSONRequestBody defines body for PostAlerts for application/json ContentType.
type PostAlertsJSONRequestBody PostAlertsJSONBody

// PutAlertsAlertIdJSONRequestBody defines body for PutAlertsAlertId for application/json ContentType.
type PutAlertsAlertIdJSONRequestBody PutAlertsAlertIdJSONBody